Update For Audit
ME-02 is Is fixed by using storage instead of instance on unbounded data

MI-01 is resolved by requiring a quote upon deposit which verifies the the user will get the expected quote, if the two quotes differs the deposit will fail

EN-02 is not implemented by design, there it is no incentive in leaving the rewards in the farm contract so we do accept that whatever the amount the user withdraw it will also withdraw all pending rewards.

Additionally if called with amount 0 the withdraw function allows just to claim rewards if the user doesn't want to withdraw from the farm.